Update chips.ts
Hello,
By using this component I've been experienced one issue while using a an array object as value and a item template. The content of the chip was not displayed. If we use a property of the input object value into the template two things happen :
- This value is not checked correctly into the addItem, so duplicate case is not managed 
- The text value is not displayed correctly into the chip since we add it straight without taking care of the type of the input object

I suggest to be able to handle thoses cases by specifying the field into the existing @input field and map the object correctly

I've not tested the code but I think it would be a good start. I'll test this if your feedback is positive about the idea.